+++ NEWS.rst
@@ -6,7 +6,7 @@ of every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Added cxx_gxx64_x86 and cxx_gxx64_x64.
 
 4.37.0
 ------